requirements.txt
Flask
numpy
pandas
matplotlib
scikit-learn
pymysql
mysql-connector-python
python-dotenv
scikit-learn
sqlalchemy